refactor(loom): hoist shared markdown build + link routing into a composable
MarkdownView and ArtifactDocument carried byte-identical build()/renderContext()
and in-document link routing (the lint review flagged the duplication). Move both
into lib/markdownDoc.ts:

- useMarkdownDoc(props, onBuilt): the parse → reactive tokens/ctx pipeline with
  the out-of-order runId guard and the source watch; onBuilt fires once the fresh
  tree has painted (MarkdownView emits `rendered`; ArtifactDocument re-locates its
  comment threads).
- routeDocLink(e, router, body): `#`-anchor scroll + smartdoc data-issue/
  data-artifact routing, returning whether it handled the click — so
  ArtifactDocument falls through to its caret hit-test only on a non-link click.

No behaviour change: full e2e green (71), tsc clean.
